article votes summedUserStance is updating

+++ client/src/apollo/resolvers.js
@@ -17,7 +17,7 @@ export const DELETE_ARTICLE = gql`
 
 // We should tweak the output, but its not that important
 export const UPDATE_ARTICLE_VOTES = gql`
-  mutation ($_id: String!, $votes: VoteInput, $userStance: Int) {
+  mutation ($_id: String!, $votes: VoteInput, $userStance: Float) {
     updateArticleVotes(_id: $_id, votes: $votes, userStance: $userStance) {
       _id
       url
